+++ infovault/lib/utils/constants.dart
@@ -1,7 +1,8 @@
 class AppConstants {
   static const int clipboardClearSeconds = 60;
   static const int autoLockMinutes = 3;
-  static const int pbkdf2Iterations = 100000;
+  static const int pbkdf2Iterations = 10000;
+  static const int legacyPbkdf2Iterations = 100000;
   static const int maxFailedAttempts = 5;
   static const int lockoutDurationMinutes = 5;
   static const int maxPhotosPerItem = 3;