+++ infovault/windows/runner/runner_window.cpp
@@ -1,1 +1,0 @@
-#include "runner_window.h"